Store hashes from non-v0 CIDs in storethehash
Currently, go-storethehash can only store keys that are CIDs.  So, if a multihash is from a non-v0 CID this causes an error when trying to transform key bytes back into a CID (since the keybytes were a multihash and not a VID).

This PR fixes this by always creating a v1 CID from the multihash before storing it into storethehash.

Added unit test to test storing hash from v1 CID.