Make a Metadata type that contains ProtocolID and Data (#27)
* Make a Metadata type that contains ProtocolID and Data

There is now a Metadata type which makes it explicit that metadata is composed of protocol id and data.  The indexer.Value has a MetadataBytes member, which is the serialized form of a Metadata instance.  This is because metadata is opaque to the indexer, and the indexer should never need to deal with the components of metadata.

* Use multicodec.Code as type for protocol id