Fix storethehash regression
The storethehash valuestore expects the first n bits of the multihash digest to be hash data (even distribution of values).  Adding a fixed value prefix caused a tragic performance regression.  The prefix was used to differentiate index keys from metadata keys.  This change fixes the problem by putting the identifier value at the end of the hash data.