Make value serialization format pluggable and add binary format
The indexer value serliazation format was hard-coded to JSON. This
forced users to use JSON, which offers good human readability but does
not offer the best performance characteristic should a user want to
trade readability with IO performance for example.

Make the serialization mechanism configurable by introducing
`ValueSerde` interface. Wrap the existing serialization functionality to
an implementation of the interface, called `JsonValueSerde` and set it
as the default serialization format for backward compatibility.

Considering the majority of data being stored by indexer engine is
already in binary format, introduce `BinaryValueSerde`.

Add a benchmark that compares the performance of both implementations.