Gracefully migrate from JSON to Binary value encoding by default
Define a new value codec that always serializes to binary but
understands both JSON and Binary format on deserialization. The idea is
to gracefully and opportunistically migrate  codec from JSON to the more
efficient binary format.

Use the newly defined codec as the default value codec for stores that
support configuring it: pogreb and sth.

Fixes #84