Fix value generator satting all values the same
Since the random value generator reuses and overwrites the generated slice, all generated slices get the value of the last generated slice. In other words, if I generate 10 values, and then examine all 10, they will be identical having the value of the last generated slice.

This PR fixes that by not reusing the slice where the values are written.