Handle merge of already serialized value-keys
Pebble documentation suggests that `Merger` is used by `.Merge` calls on
a given key value. However, Pebble uses merger on iteration over values
too. Internally, `Get` operations use iterators for example. This means
values that are already serialized may be passed to the merger
implementation during in-memory traversal.

The previous implementation of value-keys merger did not take this into
account and as a result it meant that `Get` failed to find value-keys
associated to a multihash since they ended up being double wrapped by
the merger value-keys marshaller.

The changes here fixes this issue by checking explicitly that a give
value-key has the expected key prefix before proceeding with the
remaining merge logic. As a precaution, the merger is also modified to
fall back on the default Pebble merger when the key is not of kind
multihash-key. Further, the benchmarks are modified to explicitly assert
that the inserted values are indeed found for each entry.

The `Get` implementation in Pebble store is also modified to copy the
content returned by the backing store to avoid slice invalidation and
runtime invalid reference errors captured in #80.

Unrelated to the lookup issue, the changes here also fix typos in
benchmarks where workload does not match the benchmark name. Further,
the common tests are updated to do a flush as suggested in #93.

Fixes #94, #80.