Gracefully recover from nested value-key marshalling
The previous behaviour of value-key merger in certain scenarios may have
been resulted in value-keys being marshalled repeatedly. This was the
underlying cause of the issue captured in #94 and corrected in #95.

The work here adds to the corrections by gracefully fixing such records
by unmarshalling them recursively until we reach values with the
expected key prefix, and de-duplicating them as necessary. The fix is
done opportunistically, in that whenever such values are detected during
reads and writes they are fixed.

Additional tests are added to simulate nested marshalling of value-keys
and assert that the resulting merge is as expected for both merge-older
and merge-newer cases.