Introduce `Stats` API to expose valustore statistics
Expose an API that, when supported by the backing store, it returns the
number of unique multihashes stored.

Implement stats calculation for pebble as an estimate of the number of
entries in SST files when the key range falls with the multihash key
start/end.

Relates to: https://github.com/filecoin-project/storetheindex/issues/982