Enable key-sharded writes to dhstore (#160)
* Enable key-sharded writes to dhstore

If the option WithDHKeyShard(true) is given, then enable key-sharded writes to DHStore. When key sharding is enabled, a shard key is included with each metadata put request, and a seperate merge request with a shard key is sent for each multihash. Context deletes also include a key-shard.

Fix Get so that DHStore works when not using a result cache.

* Get ignores bad multihash
* Set batch size if given size is > 0
* Allow no valuestore if dhstore URL is provided
* dh key sharding is off by default
* Send sharded merge requests concurrently
* Option to configure dh shard concurrency
* Log elapsed time for all merges